Moving signal receiver code
